refactor: LottoRandom 추상화
